+++ analyze_results/analyze_kappa.py
@@ -1,35 +1,7 @@
-from typing import List
-
-from FleissKappa.fleiss import fleissKappa
 from scipy.stats import kendalltau
 
-from extract_results import get_aggregated_classification, get_rater_classifications
-
-
-def create_binary_scale(classification_matrix: List[List[int]]):
-    binary_classification_matrix: List[List[int]] = list()
-
-    harmless_count = 0
-    offensive_count = 0
-
-    for discrete_classification in classification_matrix:
-        binary_classification = [0, 0]
-        for i, value in enumerate(discrete_classification):
-            if i + 1 <= len(discrete_classification) / 2:
-                binary_classification[0] += value
-            else:
-                binary_classification[1] += value
-
-        if binary_classification[0] >= binary_classification[1]:
-            harmless_count += 1
-        else:
-            offensive_count += 1
-        binary_classification_matrix.append(binary_classification)
-
-    print(f"Harmless: {harmless_count} Offensive: {offensive_count} "
-          f"Percent Offensive: {offensive_count / (offensive_count + harmless_count)}")
-    return binary_classification_matrix
-
+from FleissKappa.fleiss import fleissKappa
+from extract_results import get_aggregated_classification, get_rater_classifications, create_binary_scale
 
 if __name__ == '__main__':
     db_files = [
